feat: wallet connect fence
